refactor

+++ cmd/networker.go
@@ -9,7 +9,7 @@ import (
 // Networker is the root cmd.
 var Networker = &cobra.Command{
 	Use:  "networker",
-	Long: "networker is an easy to use networking CLI tool",
+	Short: "networker is an easy to use networking CLI tool",
 	Run: func(cmd *cobra.Command, args []string) {
 		if err := cmd.Usage(); err != nil {
 			log.Fatalf("failed to print usage - %v\n", err)
